#234
Adjusted the virtualization for the ComboBox. Cleaned it up a bit.
