+++ frontend/components/reviews/ui/mobile-select-list.tsx
@@ -2,7 +2,6 @@ import {Fragment} from 'react'
 import {Listbox, Transition} from '@headlessui/react'
 import {CheckIcon, SelectorIcon} from '@heroicons/react/solid'
 import {Options} from '@/util/interfaces'
-import { RowVirtualizerVariable } from '@/util/virtualizer'
 
 interface ComponentProps {
 	name: string
@@ -43,10 +42,33 @@ export default function MobileSelectList({
 							leaveTo="opacity-0"
 						>
 							<Listbox.Options className="absolute z-10 mt-1 max-h-60 w-full overflow-auto rounded-md bg-white py-1 text-base shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none sm:text-sm">
-							<RowVirtualizerVariable 
-								rows={options ?? []} 
-								state={state}
-							/>
+								{options.map((option) => (
+									<Listbox.Option
+										key={option.id}
+										className={({active}) =>
+											`relative cursor-default select-none py-2 pl-10 pr-4 ${
+												active ? 'bg-amber-100 text-amber-900' : 'text-gray-900'
+											}`
+										}
+										value={option}
+									>
+										<span
+											className={`block truncate ${
+												option.name === state?.name
+													? 'font-medium'
+													: 'font-normal'
+											}`}
+										>
+											{option.name}
+										</span>
+
+										{option.name === state?.name ? (
+											<span className="absolute inset-y-0 left-0 flex items-center pl-3 text-amber-600">
+												<CheckIcon className="h-5 w-5" aria-hidden="true" />
+											</span>
+										) : null}
+									</Listbox.Option>
+								))}
 							</Listbox.Options>
 						</Transition>
 					</div>
