+++ frontend/components/reviews/ui/select-list.tsx
@@ -2,7 +2,6 @@ import React, {Fragment} from 'react'
 import {CheckIcon, SelectorIcon} from '@heroicons/react/solid'
 import {Listbox, Transition} from '@headlessui/react'
 import {Options} from '@/util/interfaces'
-import { RowVirtualizerVariable } from '@/util/virtualizer'
 
 interface ComponentProps {
 	name: string
@@ -39,10 +38,31 @@ export default function SelectList({
 					leaveTo="opacity-0"
 				>
 					<Listbox.Options className="absolute mt-1 max-h-60 w-60 overflow-auto rounded-md bg-white py-1 text-base shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none sm:text-sm">
-					<RowVirtualizerVariable 
-						rows={options ?? []} 
-						state={state}
-					/>
+						{options.map((option) => (
+							<Listbox.Option
+								key={option.id}
+								className={({active}) =>
+									`relative cursor-default select-none py-2 pl-10 pr-4 ${
+										active ? 'bg-amber-100 text-amber-900' : 'text-gray-900'
+									}`
+								}
+								value={option}
+							>
+								<span
+									className={`block truncate ${
+										option.name === state?.name ? 'font-medium' : 'font-normal'
+									}`}
+								>
+									{option.name}
+								</span>
+
+								{option.name === state?.name ? (
+									<span className="absolute inset-y-0 left-0 flex items-center pl-3 text-amber-600">
+										<CheckIcon className="h-5 w-5" aria-hidden="true" />
+									</span>
+								) : null}
+							</Listbox.Option>
+						))}
 					</Listbox.Options>
 				</Transition>
 			</div>
